src/main/java/com/telstra/controller/ServiceErrorController.java
package com.telstra.controller;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

import com.telstra.exceptions.ErrorResponse;
import com.telstra.exceptions.InvalidInputValidationException;

@ControllerAdvice
public class ServiceErrorController {
 
	@ExceptionHandler(InvalidInputValidationException.class)
	public ResponseEntity<ErrorResponse> exceptionHandler(Exception ex) {
		ErrorResponse error = new ErrorResponse();
		error.setErrorCode(HttpStatus.PRECONDITION_FAILED.value());
		error.setMessage(ex.getMessage());
		return new ResponseEntity<ErrorResponse>(error, HttpStatus.NOT_FOUND);
	}
}
